refactor web and state manager wiring for PR1
Split core HTTP handlers and shared state-manager types/config into dedicated files to reduce main.go complexity while preserving behavior.

Made-with: Cursor
